backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Dmitry Nosov <millttt7@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.12"
pydantic-settings = ">=2.8.1,<3.0.0"
alembic = ">=1.15.2,<2.0.0"
sqlalchemy = ">=2.0.40,<3.0.0"
fastapi = { version = ">=0.115.12,<0.116.0", extras = ["standard"] }
python-multipart = ">=0.0.20,<0.0.21"
aiobotocore = ">=2.21.1,<3.0.0"
asyncpg = ">=0.30.0,<0.31.0"
types-aiobotocore = { version = ">=2.21.1,<3.0.0", extras = ["s3"] }
pyjwt = ">=2.10.1,<3.0.0"
cryptography = ">=44.0.2,<45.0.0"
psycopg2-binary = "^2.9.10"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ruff = "^0.11.4"


[tool.pyright]
venvPath = "."
venv = ".venv"
pythonVersion = "3.12"

[tool.ruff]
line-length = 120
